# contracts/openapi/mindspace.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_destiny_avatar_api_v1_destiny_avatars_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_destiny_avatar_api_v1_destiny_avatars_post",
        "type": "object"
      },
...
      "DestinySeed": {
        "properties": {
          "adult_character": {
            "default": true,
            "title": "Adult Character",
            "type": "boolean"
          },
          "ai_gender": {
            "default": "不指定",
            "enum": [
              "女",
              "男",
              "不指定"
            ],
            "title": "Ai Gender",
            "type": "string"
          },
          "ai_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Ai Name",
            "type": "string"
          },
          "appearance_expectation": {
            "default": "",
            "maxLength": 1200,
            "title": "Appearance Expectation",
            "type": "string"
          },
          "avatar": {
            "additionalProperties": true,
            "title": "Avatar",
            "type": "object"
          },
          "character_expectation": {
            "maxLength": 2400,
            "minLength": 1,
            "title": "Character Expectation",
            "type": "string"
          },
          "relationship": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Relationship",
            "type": "string"
          },
          "relationship_context": {
            "default": "",
            "maxLength": 2400,
            "title": "Relationship Context",
            "type": "string"
          },
          "user_alias": {
            "default": "",
            "maxLength": 80,
            "title": "User Alias",
            "type": "string"
          },
          "user_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "ai_name",
          "user_name",
          "relationship",
          "character_expectation"
        ],
        "title": "DestinySeed",
        "type": "object"
      },
      "DestinySelectionRequest": {
        "properties": {
          "card_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Card Id",
            "type": "string"
          },
          "expected_revision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          }
        },
        "required": [
          "card_id"
        ],
        "title": "DestinySelectionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mindspace Graph",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/destiny/avatars": {
      "post": {
        "description": "Store a journey-local avatar without touching global avatar settings.",
        "operationId": "upload_destiny_avatar_api_v1_destiny_avatars_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_destiny_avatar_api_v1_destiny_avatars_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Destiny Avatar"
      }
    },
    "/api/v1/destiny/avatars/{filename}": {
      "delete": {
        "operationId": "discard_destiny_avatar_api_v1_destiny_avatars__filename__delete",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Destiny Avatar"
      }
    },
...
    "/api/v1/destiny/journeys": {
      "post": {
        "operationId": "create_destiny_journey_api_v1_destiny_journeys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinySeed"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Destiny Journey"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}": {
      "get": {
        "operationId": "get_destiny_journey_api_v1_destiny_journeys__journey_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Destiny Journey"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/archetypes": {
      "post": {
        "operationId": "generate_destiny_archetypes_api_v1_destiny_journeys__journey_id__archetypes_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "use_default",
            "required": false,
            "schema": {
              "default": false,
              "title": "Use Default",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Destiny Archetypes"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/cards": {
      "post": {
        "operationId": "generate_destiny_cards_api_v1_destiny_journeys__journey_id__cards_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "use_default",
            "required": false,
            "schema": {
              "default": false,
              "title": "Use Default",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Destiny Cards"
      }
    },
...
    "/api/v1/destiny/journeys/{journey_id}/commit": {
      "post": {
        "operationId": "commit_destiny_journey_api_v1_destiny_journeys__journey_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Destiny Journey"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/rewind/archetypes": {
      "post": {
        "operationId": "rewind_destiny_archetypes_api_v1_destiny_journeys__journey_id__rewind_archetypes_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rewind Destiny Archetypes"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/selections": {
      "delete": {
        "operationId": "clear_destiny_selections_api_v1_destiny_journeys__journey_id__selections_delete",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Destiny Selections"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/selections/{slot_id}": {
      "delete": {
        "operationId": "unselect_destiny_card_api_v1_destiny_journeys__journey_id__selections__slot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "title": "Slot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unselect Destiny Card"
      },
      "put": {
        "operationId": "select_destiny_card_api_v1_destiny_journeys__journey_id__selections__slot_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "title": "Slot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinySelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Destiny Card"
      }
    },
    "/api/v1/destiny/journeys/{journey_id}/synthesize": {
      "post": {
        "operationId": "synthesize_destiny_journey_api_v1_destiny_journeys__journey_id__synthesize_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Synthesize Destiny Journey"
      }
    },
...
  }
}